Feat: Improve code implementation (FM-484)
